add support for "length"
xcb-proto add a new element in c36dde3f4535 ("Add element to specify
expression that defines length of a struct"); let's support parsing it.

Note that we hard-code the type here since we require expressions to have a
type, though they are not strictly necessary here since length does not
appear on the wire. See complete discussion here:

https://lists.freedesktop.org/archives/xcb/2021-September/011385.html

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>